Formatting tests for all applications
