+++ raspeedi/tests/tests_raspeedi_cmds.py
