--- squalaetp/tests/__init__.py
@@ -1,4 +1,0 @@
-from .corvet_views import *
-from .corvet_cmds import *
-from .xelon_views import *
-from .xelon_cmds import *
