+++ squalaetp/tests/tests_corvet_selenium_views.py
@@ -1,0 +1,73 @@
+from django.test import LiveServerTestCase
+from django.contrib.auth import get_user_model
+
+from selenium import webdriver
+from selenium.webdriver.firefox.options import Options
+
+from squalaetp.models import Corvet
+
+User = get_user_model()
+
+
+class CorvetSeleniumTestCase(LiveServerTestCase):
+
+    def setUp(self):
+        options = Options()
+        options.add_argument('-headless')
+        self.driver = webdriver.Firefox(firefox_options=options)
+        self.driver.implicitly_wait(30)
+        super(CorvetSeleniumTestCase, self).setUp()
+        User.objects.create_user(username='toto', email='toto@bibi.com', password='totopassword')
+        self.data = (
+            '<?xml version="1.0" encoding="UTF-8"?><MESSAGE><ENTETE><EMETTEUR>CLARION_PROD</EMETTEUR></ENTETE>'
+            '<VEHICULE Existe="O">'
+            '<DONNEES_VEHICULE><WMI>VF3</WMI><VDS>ABCDEF</VDS><VIS>12345678</VIS>'
+            '<TRANSMISSION>0R</TRANSMISSION></DONNEES_VEHICULE>'
+            '<LISTE_ATTRIBUTS><ATTRIBUT>DAT24</ATTRIBUT><ATTRIBUT>GG805</ATTRIBUT></LISTE_ATTRIBUTS>'
+            '<LISTE_ORGANES><ORGANE>10JBCJ3028478</ORGANE><ORGANE>20DS850837512</ORGANE></LISTE_ORGANES>'
+            '<LISTE_ELECTRONIQUES><ELECTRONIQUE>14A9666571380</ELECTRONIQUE>'
+            '<ELECTRONIQUE>P4A9666220599</ELECTRONIQUE></LISTE_ELECTRONIQUES>'
+            '</VEHICULE></MESSAGE>'
+        )
+        self.vin = 'VF3ABCDEF12345678'
+
+    def tearDown(self):
+        self.driver.quit()
+        super(CorvetSeleniumTestCase, self).tearDown()
+
+    def test_corvet_table_page_is_connected(self):
+        driver = self.driver
+
+        # Creating session cookie for to access Raspeedi insert form
+        self.client.login(username='toto', password='totopassword')
+        cookie = self.client.cookies['sessionid']
+        driver.get(self.live_server_url + '/squalaetp/corvet/')
+        driver.add_cookie({'name': 'sessionid', 'value': cookie.value, 'secure': False, 'path': '/'})
+        driver.refresh()
+        driver.get(self.live_server_url + '/squalaetp/corvet/')
+
+        self.assertEqual(driver.current_url, self.live_server_url + '/squalaetp/corvet/')
+
+    def test_corvet_insert_is_valid(self):
+        driver = self.driver
+        old_corvet = Corvet.objects.count()
+
+        # Creating session cookie for to access Raspeedi insert form
+        self.client.login(username='toto', password='totopassword')
+        cookie = self.client.cookies['sessionid']
+        driver.get(self.live_server_url + '/squalaetp/corvet/insert/')
+        driver.add_cookie({'name': 'sessionid', 'value': cookie.value, 'secure': False, 'path': '/'})
+        driver.refresh()
+        driver.get(self.live_server_url + '/squalaetp/corvet/insert/')
+
+        # Inserting values into the form
+        vin = driver.find_element_by_name('vin')
+        xml_data = driver.find_element_by_name('xml_data')
+        submit = driver.find_element_by_name('btn_corvet_insert')
+        vin.send_keys(self.vin)
+        xml_data.send_keys(self.data)
+        submit.click()
+
+        new_corvet = Corvet.objects.count()
+        self.assertEqual(new_corvet, old_corvet + 1)
+        self.assertEqual(driver.current_url, self.live_server_url + '/squalaetp/corvet/insert/')
